requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
openai>=1.0.0
anthropic>=0.18.0
